+++ CHANGELOG.md
@@ -423,6 +423,10 @@
   [p4checo](https://github.com/p4checo)
   [#5965](https://github.com/realm/SwiftLint/issues/5965)
 
+* Add opt-in `unneeded_throws_rethrows` rule that triggers when declarations
+  marked `throws`/`rethrows` never actually throw or call any throwing code.  
+  [Tony Ngo](https://github.com/tonyskansf)
+
 ### Bug Fixes
 
 * Fix `no_extension_access_modifier` rule incorrectly triggering for `nonisolated extension`.